django_app/requirements.txt
# base
django
django-environ
psycopg2-binary
gunicorn

# dev
django-debug-toolbar
django-extensions
requests
tqdm
pytest-django
factory-boy
django-browser-reload